cputopology: Unify online/offline cpu code
The online/offline cpu code can be unified which leads to simplified code.

Add a cpustatus function that takes cpu number and the cpu state (0 for
offline and 1 for online) as arguments.  Modify the code to make use of
the cpustatus function.

There are no functional changes with this patch.

[Also while I'm at it, make the cputopology file executable.]

Signed-off-by: Prarit Bhargava <prarit@redhat.com>